+++ setup.py
@@ -4,19 +4,19 @@ from setuptools import setup
 
 setup(
     name="autoupgrade-prima",
-    version="0.5.1",
+    version="0.6.0",
     author="Walter Purcaro",
     author_email="vuolter@gmail.com",
     description="Automatic upgrade of PyPI packages",
-    long_description=open('README.rst').read(),
-    keywords=['autoupgrade', 'pip-upgrade', 'pip'],
-    packages=['autoupgrade'],
+    long_description=open("README.rst").read(),
+    keywords=["autoupgrade", "pip-upgrade", "pip"],
+    packages=["autoupgrade"],
     include_package_data=True,
     url="https://github.com/primait/autoupgrade",
     download_url="https://github.com/primait/autoupgrade/releases",
-    install_requires=['requests'],
-    obsoletes=['autoupgrade'],
-    license='MIT License',
+    install_requires=["requests"],
+    obsoletes=["autoupgrade"],
+    license="MIT License",
     zip_safe=True,
     classifiers=[
         "Development Status :: 5 - Production/Stable",
@@ -29,6 +29,6 @@ setup(
         "Programming Language :: Python :: 2",
         "Programming Language :: Python :: 3",
         "Topic :: System :: Software Distribution",
-        "Topic :: Utilities"
-    ]
+        "Topic :: Utilities",
+    ],
 )
